content_creation_agency/requirements.txt
agency-swarm==0.4.2
python-dotenv>=0.19.0
openai>=1.0.0
pytrends>=4.9.0
requests>=2.31.0
tavily-python>=0.1.9
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
nltk>=3.8.1
google-api-python-client>=2.0.0
google-genai>=1.10.0
textblob>=0.17.1
flask>=3.0.0
flask-cors>=4.0.0 